requirements.txt
# Core Framework
streamlit==1.29.0
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# AI/ML Libraries
transformers==4.36.0
torch==2.1.0
sentence-transformers==2.2.2

# OCR
pytesseract==0.3.10
pdf2image==1.16.3

# Computer Vision
opencv-python==4.9.0.80
Pillow==10.2.0

# NLP
spacy==3.7.2

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.3

# Database
sqlalchemy==2.0.25

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pandas==2.1.4
numpy==1.26.3

# Testing
pytest==7.4.3
httpx==0.26.0
